+++ src/main/java/com/biwaby/financialtracker/repository/LimitRepository.java
@@ -3,5 +3,5 @@ package com.biwaby.financialtracker.repository;
 import com.biwaby.financialtracker.entity.Limit;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LimitRepository extends JpaRepository<Limit, Integer> {
+public interface LimitRepository extends JpaRepository<Limit, Long> {
 }
