+++ src/main/java/com/biwaby/financialtracker/repository/RoleRepository.java
@@ -3,5 +3,5 @@ package com.biwaby.financialtracker.repository;
 import com.biwaby.financialtracker.entity.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoleRepository extends JpaRepository<Role, Integer> {
+public interface RoleRepository extends JpaRepository<Role, Long> {
 }
